Update packaging
- refined view
- added manual creation
- bug fix

+++ Data/Module3/P2-5/Interfaces/IPackagingProfileGateway.cs
@@ -5,7 +5,7 @@ namespace ProRental.Data.Module3.P2_5.Interfaces;
 
 public interface IPackagingProfileGateway
 {
-    void Save(int orderId, float volume, string fragilityLevel);
+    void Save(int orderId, double volume, string fragilityLevel);
     Packagingprofile FindByProfileId(int profileId);
     Packagingprofile FindByOrderId(int orderId);
     List<PackagingFootprintProfileDto> GetAllFootprintProfiles();
